Move redundant 1D-to-multiD array creation code into one function, with same name as soon (hopefully) to be committed construct_md_array function that will be in pg7.4. The function is wrapped in #if(CATALOG_VERSION_NO < 200303091), and the actual catalog version number introducing the new function will have to replace the current placeholder later.
Add new reload_plr_modules() function. This allows a forced reload of R
functions from the plr_modules table, mostly so that a PL/R function can
be used as part of the install script (see below) and yet the plr_modules
loading functionality can still be tested in the regression test.

Removed some overly-anal typechecking in the array_* functions. This is to be
consistent with the versions submitted for inclusion in the backend. We'll
depend on users not to abuse the functions when they define them. This should
have a relatively significant impact on speed.

Added a PL/R function to the install script that allows conditional creation
of the C functions. This is to avoid creation of the PL/R versions of the
array_* functions if they are already in the backend (as they will be in
pg7.4)

Minor changes to regression script and expected output to deal with the above.